update name, app.py -> main.py, main.py -> run_cli.py(run command line interface)
